Replace content_for backing with OutputFlow implementation
Rails' content_for delegates its storage to @view_flow,
an instance of ActiveSupport::OutputFlow. It's really just a
wrapper around a hash of ActiveSupport::SafeBuffers.

https://github.com/rails/rails/blob/b8bd24d93825684e2311d085266cf66b4fb45368/actionview/lib/action_view/flows.rb#L10

I think we're better off inlining the indirection here and take on some
of content_for's implementation.

This also means we can skip the costly SecureRandom.uuid, and additionally
make Partial have as little a memory impact as possible so it's not as
dangerous to allocate an extra one.