Store t prefix on the stack
For the array storage we only care about the last value, which means
that we're able to just store the prefix in a single string variable.

We can still push the prefix if we're nesting further, or reset it to
not leak if we're performing a call without a block.

We can then pop/revert the prefix by resetting the variable.