Base auto-capture logic on yield calls in template source
Currently, we're basing our auto-capture logic on whether there's a
single argument passed to render, e.g.:

```ruby
render "card" do |partial|
  partial.yield :title, t(".title")
end
```

But if users want to pass extra content via an explicit `<%= yield something %>`,
we don't support that.

This changes our logic to look for `<%= yield something %>` calls in the
template source and not auto-capture in that case. We will pass the partial
on as an extra argument in case callers want to expose that as a named block
argument.

We'll still auto-capture in case a template has non-capturing yields like
`<%= yield :title %>`.