Use load hooks for monkey_patch loading
`ActiveSupport.on_load(:action_view)` is technically how we're supposed
to extend ActionView::Base and other Action View classes, but we weren't
doing that for our monkey patches.

We could add a `ActiveSupport.on_load(:action_view)` load hook for each
extension, but it feels cleaner to let the top level file handle that
set up.

This then also moves the helper require to the same place for parity.