Only call `concat` for `String` descendants
When writing content, only concatenate "argument" content when its a
descendant of [String][] (this includes instances of
[ActiveSupport::SafeBuffer][] created by calls to `#html_safe`).

If content is `nil` or an object, capture content from the `&block`.

The conditional guard **intentionally** compares `#class` and `String`
with the [<=][] operator instead of the [Object#is_a?][] predicate so
that object instances created by [with_options][] can pass through the
conditional without invoking `is_a?` with the incorrect number of
arguments.

[String]: https://ruby-doc.org/core/String.html
[ActiveSupport::SafeBuffer]: https://edgeapi.rubyonrails.org/classes/ActiveSupport/SafeBuffer.html
[<=]: https://ruby-doc.org/core-3.1.2/Module.html#method-i-3C-3D
[Object#is_a?]: https://ruby-doc.org/core-3.1.2/Object.html#method-i-is_a-3F
[with_options]: https://edgeapi.rubyonrails.org/classes/Object.html#method-i-with_options